Reorganize new logic back into analysis.go
